+++ contrib_bots/bots/DefineBot/docs.md
