+++ contrib_bots/bots/EncryptBot/docs.md
