+++ contrib_bots/bots/FourSquareBot/settings.ini
