+++ contrib_bots/bots/John/var/jokes.json
