+++ contrib_bots/bots/define_bot.py
