+++ contrib_bots/bots/xkcd/xkcd-wrong-command.png
