+++ contrib_bots/run.py
@@ -12,19 +12,18 @@ sys.path.insert(0, our_dir)
 
 from bot_lib import run_message_handler_for_bot
 
-def get_lib_module(lib_fn):
-    lib_fn = os.path.abspath(lib_fn)
-    if not os.path.dirname(lib_fn).startswith(os.path.join(our_dir, 'lib')):
-        print('Sorry, we will only import code from contrib_bots/lib.')
+def get_lib_module(bots_fn):
+    bots_fn = os.path.abspath(bots_fn)
+    if not os.path.dirname(bots_fn).startswith(os.path.join(our_dir, 'bots')):
+        print('Sorry, we will only import code from contrib_bots/bots.')
         sys.exit(1)
 
-    if not lib_fn.endswith('.py'):
+    if not bots_fn.endswith('.py'):
         print('Please use a .py extension for library files.')
         sys.exit(1)
-
-    sys.path.append('lib')
-    base_lib_fn = os.path.basename(os.path.splitext(lib_fn)[0])
-    module_name = 'lib.' + base_lib_fn
+    base_bots_fn = os.path.basename(os.path.splitext(bots_fn)[0])
+    sys.path.append('bots/{}'.format(base_bots_fn))
+    module_name = base_bots_fn
     module = importlib.import_module(module_name)
     return module
 
@@ -58,7 +57,7 @@ def run():
         print('You must specify a library!')
         sys.exit(1)
 
-    lib_module = get_lib_module(lib_fn=args[0])
+    lib_module = get_lib_module(bots_fn=args[0])
 
     if not options.quiet:
         logging.basicConfig(stream=sys.stdout, level=logging.INFO)
